requirements.txt
django
djangorestframework
pillow
GitPython